refactor: split models.py into organized modules
Split 403-line models.py into 4 modules (providers, agents, rag, tasks) for better organization. All imports remain backward compatible.
