+++ bolna/models/__init__.py
@@ -1,0 +1,113 @@
+# Re-export all models for backward compatibility
+# This ensures that imports like "from bolna.models import AgentModel" continue to work
+
+# Provider configs
+from .providers import (
+    PollyConfig,
+    ElevenLabsConfig,
+    OpenAIConfig,
+    DeepgramConfig,
+    CartesiaConfig,
+    RimeConfig,
+    SmallestConfig,
+    SarvamConfig,
+    AzureConfig,
+)
+
+# RAG and Vector Store configs
+from .rag import (
+    MongoDBProviderConfig,
+    RerankerConfig,
+    LanceDBProviderConfig,
+    VectorStore,
+    GraphNodeRAGConfig,
+)
+
+# Agent models
+from .agents import (
+    Route,
+    Routes,
+    Llm,
+    SimpleLlmAgent,
+    Node,
+    Edge,
+    LlmAgentGraph,
+    GraphEdge,
+    GraphNode,
+    GraphAgentConfig,
+    KnowledgeAgentConfig,
+    AgentRouteConfig,
+    MultiAgent,
+    KnowledgebaseAgent,
+    LlmAgent,
+)
+
+# Tasks, Tools, IO, and Conversation configs
+from .tasks import (
+    AGENT_WELCOME_MESSAGE,
+    validate_attribute,
+    Transcriber,
+    Synthesizer,
+    IOModel,
+    ToolFunction,
+    ToolDescription,
+    ToolDescriptionLegacy,
+    APIParams,
+    ToolModel,
+    ToolsConfig,
+    ToolsChainModel,
+    ConversationConfig,
+    Task,
+    AgentModel,
+)
+
+__all__ = [
+    # Provider configs (9)
+    "PollyConfig",
+    "ElevenLabsConfig",
+    "OpenAIConfig",
+    "DeepgramConfig",
+    "CartesiaConfig",
+    "RimeConfig",
+    "SmallestConfig",
+    "SarvamConfig",
+    "AzureConfig",
+    # RAG configs (5)
+    "MongoDBProviderConfig",
+    "RerankerConfig",
+    "LanceDBProviderConfig",
+    "VectorStore",
+    "GraphNodeRAGConfig",
+    # Agent models (15)
+    "Route",
+    "Routes",
+    "Llm",
+    "SimpleLlmAgent",
+    "Node",
+    "Edge",
+    "LlmAgentGraph",
+    "GraphEdge",
+    "GraphNode",
+    "GraphAgentConfig",
+    "KnowledgeAgentConfig",
+    "AgentRouteConfig",
+    "MultiAgent",
+    "KnowledgebaseAgent",
+    "LlmAgent",
+    # Tasks and configs (15)
+    "AGENT_WELCOME_MESSAGE",
+    "validate_attribute",
+    "Transcriber",
+    "Synthesizer",
+    "IOModel",
+    "ToolFunction",
+    "ToolDescription",
+    "ToolDescriptionLegacy",
+    "APIParams",
+    "ToolModel",
+    "ToolsConfig",
+    "ToolsChainModel",
+    "ConversationConfig",
+    "Task",
+    "AgentModel",
+]
